+++ src/components/Form/CheckboxInput/Checkbox.tsx
@@ -1,6 +1,6 @@
 import React, { ChangeEvent } from 'react';
 // eslint-disable-next-line import/no-cycle
-import { ComponentBaseInterface } from '../Form';
+import { ComponentBaseInterface } from '../../../data/types/type';
 
 interface InterfaceCheckbox extends ComponentBaseInterface {
   checkboxChangeHandler: (event: ChangeEvent<HTMLInputElement>) => void;
