+++ src/components/Form/index.ts
@@ -1,7 +1,1 @@
-export { Birthday } from './BirthdayInput';
-export { Checkbox } from './CheckboxInput';
 export { Form } from './Form';
-export { Name } from './NameInput';
-export { Option } from './OptionInput';
-export { Radio } from './RadioInput';
-export { Surname } from './SurnameInput';
