+++ src/components/MainComponent/index.ts
@@ -1,3 +1,1 @@
-export * from './Content';
 export { MainComponent } from './MainComponent';
-export * from './SearchBar';
